fix(spark): show the unified balance on the Send/Receive cards
The cards showed info.balance_sats — the BTC-only balance (e.g. "214") — but with Stable Balance on the SDK auto-sweeps BTC into USDB, so that reading badly understates a funded wallet. The Home card already shows the unified total (BTC + USDB-as-sats); the cards now match it.

Extract that computation into a shared unified_spark_balance_sats helper (also adopted by GlobalHome, so Home and the cards can't diverge), fold the Stable Balance snapshot into the get_info fetch, and compute the unified sats in the panel handlers (which have the cache's BTC/USD reference price). Also append the SATS/BTC unit suffix, matching the Home card.

Corrects the BTC-only choice from the previous card-balance commit.
