+++ coincube-gui/src/app/state/spark/mod.rs
@@ -20,11 +20,13 @@ pub use send::{SparkSend, SparkSendPhase};
 pub use settings::SparkSettings;
 pub use transactions::SparkTransactions;
 
+use std::convert::TryInto;
 use std::sync::Arc;
 
-use coincube_spark_protocol::PaymentSummary;
+use coincube_spark_protocol::{PaymentSummary, StableBalanceSnapshot};
 use iced::Task;
 
+use crate::app::cache::Cache;
 use crate::app::message::Message;
 use crate::app::wallets::SparkBackend;
 
@@ -50,19 +52,54 @@ pub(crate) fn fetch_payments_task(
     )
 }
 
-/// Fetch the Spark wallet's spendable BTC balance (sats) via `get_info`, for
-/// the two-card "YOU SEND / YOU RECEIVE" balance line. Best-effort UI polish:
-/// an error resolves to `None` so the caller keeps its last value rather than
-/// surfacing a failure. Reports the *Bitcoin* balance only, not the unified
-/// BTC + Stable Balance total — the card is labelled "Bitcoin".
+/// Fetch the Spark wallet balance via `get_info`, for the two-card
+/// "YOU SEND / YOU RECEIVE" balance line: the BTC balance plus the Stable
+/// Balance snapshot, which the caller folds into a unified sats total with
+/// [`unified_spark_balance_sats`]. Best-effort UI polish — an error resolves to
+/// `None` so the caller keeps its last value rather than surfacing a failure.
 pub(crate) fn fetch_balance_task(
     backend: Option<Arc<SparkBackend>>,
-    on_result: impl FnOnce(Option<u64>) -> Message + Send + 'static,
+    on_result: impl FnOnce(Option<(u64, Option<StableBalanceSnapshot>)>) -> Message + Send + 'static,
 ) -> Task<Message> {
     let Some(backend) = backend else {
         return Task::none();
     };
     Task::perform(async move { backend.get_info().await }, move |result| {
-        on_result(result.ok().map(|info| info.balance_sats))
+        on_result(
+            result
+                .ok()
+                .map(|info| (info.balance_sats, info.stable_balance)),
+        )
     })
+}
+
+/// The unified Spark balance in sats: the BTC balance plus any Stable Balance
+/// (USDB) holding, converted at the cache's BTC/USD reference price. Shared by
+/// the Home card and the Send/Receive card subtitles so every screen shows the
+/// same figure — Stable Balance auto-sweeps BTC into USDB, so a BTC-only
+/// reading would understate a funded wallet.
+pub(crate) fn unified_spark_balance_sats(
+    btc_sats: u64,
+    stable: Option<&StableBalanceSnapshot>,
+    cache: &Cache,
+) -> u64 {
+    // `btc_usd_price` is only set when the user's fiat preference is USD; for
+    // other fiats fall back to the user-fiat converter's per-BTC price so the
+    // holding still shows (with a small FX-spread approximation) rather than
+    // collapsing to zero. Mirrors `global_home`'s SparkBalanceUpdated handler.
+    let reference_price = cache.btc_usd_price.or_else(|| {
+        let converter: Option<crate::app::view::FiatAmountConverter> =
+            cache.fiat_price.as_ref().and_then(|p| p.try_into().ok());
+        converter.map(|c| c.price_per_btc())
+    });
+    let usdb_as_sats = stable
+        .map(|sb| {
+            crate::app::breez_spark::assets::stable_token_as_sats(
+                sb.balance,
+                sb.decimals,
+                reference_price,
+            )
+        })
+        .unwrap_or(0);
+    btc_sats.saturating_add(usdb_as_sats)
 }
