+++ coincube-gui/src/app/state/spark/send.rs
@@ -208,8 +208,9 @@ impl SparkSendTarget {
 /// Real Spark Send panel.
 pub struct SparkSend {
     backend: Option<Arc<SparkBackend>>,
-    /// The Spark wallet's spendable BTC balance (sats), shown on the YOU SEND
-    /// card. Refreshed on reload via `get_info`; `0` until the first fetch.
+    /// The Spark wallet's unified balance in sats (BTC + Stable Balance), shown
+    /// on the YOU SEND card. Refreshed on reload via `get_info`; `0` until the
+    /// first fetch.
     balance_sats: u64,
     /// Free-text destination input (BOLT11 / BIP21 / on-chain address).
     pub destination_input: String,
@@ -864,8 +865,9 @@ impl State for SparkSend {
                 Task::none()
             }
             SparkSendMessage::BalanceLoaded(balance) => {
-                if let Some(sats) = balance {
-                    self.balance_sats = sats;
+                if let Some((btc_sats, stable)) = balance {
+                    self.balance_sats =
+                        super::unified_spark_balance_sats(btc_sats, stable.as_ref(), cache);
                 }
                 Task::none()
             }
